requirements.txt
click==7.1.2
Flask==1.1.2
Flask-HTTPAuth==4.1.0
flask-marshmallow==0.13.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
PyJWT==1.7.1
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
